+++ R/snapshot.R
@@ -591,15 +591,15 @@ renv_snapshot_description <- function(path = NULL, package = NULL) {
 
 renv_snapshot_description_source <- function(dcf) {
 
+  type <- dcf[["RemoteType"]]
+  if (!is.null(type))
+    return(list(Source = renv_alias(type)))
+
   if (!is.null(dcf[["Repository"]]))
     return(list(Source = "Repository", Repository = dcf[["Repository"]]))
 
   if (!is.null(dcf[["biocViews"]]))
     return(list(Source = "Bioconductor"))
-
-  type <- dcf[["RemoteType"]]
-  if (!is.null(type))
-    return(list(Source = renv_alias(type)))
 
   package <- dcf[["Package"]]
   if (is.null(package))