+++ R/retrieve.R
@@ -209,8 +209,8 @@ renv_retrieve_git <- function(record) {
 
   renv_git_preflight()
 
-  package <- renv_tempfile("renv-git-")
-  ensure_directory(package)
+  path <- tempfile("renv-git-")
+  ensure_directory(path)
 
   template <- c(
     "cd \"${DIR}\"",
@@ -221,7 +221,7 @@ renv_retrieve_git <- function(record) {
   )
 
   data <- list(
-    DIR    = renv_path_normalize(package),
+    DIR    = renv_path_normalize(path),
     ORIGIN = record$RemoteUrl,
     REF    = record$RemoteSha %||% record$RemoteRef
   )
@@ -259,9 +259,7 @@ renv_retrieve_git <- function(record) {
     stopf(fmt, record$Package, record$RemoteUrl, status)
   }
 
-  url <- paste("file://", package, sep = "")
-  path <- renv_retrieve_path(record)
-  renv_retrieve_package(record, url, path)
+  renv_retrieve_successful(record, path)
 
 }