+++ R/tests.R
@@ -164,7 +164,7 @@ renv_tests_init_repos <- function(repos = NULL) {
   options(renv.tests.repos = c(CRAN = repos))
 
   # and update our repos option
-  fmt <- if (renv_platform_windows()) "file:%s" else "file://%s"
+  fmt <- if (renv_platform_windows()) "file:///%s" else "file://%s"
   options(
     pkgType = "source",
     repos = c(CRAN = sprintf(fmt, repos))