+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # renv 0.9.3 (UNRELEASED)
 
+* `renv::use_python("~/path/to/python")` now works as expected.
+
 * `renv` now properly expands `R_LIBS_SITE` and `R_LIBS_USER` when set within a
   startup `.Renviron` file. (#318)