+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 0.9.3 (UNRELEASED)
 
+* `renv` will now transform RSPM source URLs into binary URLs as appropriate,
+  allowing `renv` to use binary packages during restore on Linux. (#124)
+  
 * `renv` now sandboxes the system library by default on Windows.
 
 * `renv` now validates that the Xcode license has been accepted before